- move some tests to more fitting packages - import the correct maven dependencies for the arquillian tests
