--- lending-admin-server/src/main/java/info/armado/ausleihe/remote/implementation/rest/AusleiheApplication.java
@@ -1,16 +1,0 @@
-/**
- * 
- */
-package info.armado.ausleihe.remote.implementation.rest;
-
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
-
-/**
- * @author marc
- *
- */
-@ApplicationPath("/rest")
-public class AusleiheApplication extends Application {
-
-}
