+++ lending-admin-server/src/main/scala/info/armado/ausleihe/admin/remote/LendingApplication.scala
@@ -1,0 +1,12 @@
+package info.armado.ausleihe.admin.remote
+
+import javax.ws.rs.ApplicationPath
+import javax.ws.rs.core.Application
+
+
+/**
+  * @author marc
+  *
+  */
+@ApplicationPath("/rest")
+class LendingApplication extends Application
