+++ lending-admin-server/src/test/scala/info/armado/ausleihe/admin/remote/services/WebDeployment.scala
@@ -1,4 +1,4 @@
-package info.armado.ausleihe.database
+package info.armado.ausleihe.admin.remote.services
 
 import org.jboss.arquillian.container.test.api.Deployment
 import org.jboss.shrinkwrap.api.ShrinkWrap
@@ -9,7 +9,7 @@ import org.jboss.shrinkwrap.resolver.api.maven.Maven
 trait WebDeployment {
   @Deployment
   def createDeployment: WebArchive = ShrinkWrap.create(classOf[WebArchive])
-    .addPackages(true, "info.armado.ausleihe.database")
+    .addPackages(true, "info.armado.ausleihe")
     .addAsResource("test-persistence.xml", "META-INF/persistence.xml")
     .addAsResource("datasets/")
     .addAsWebInfResource(EmptyAsset.INSTANCE, "beans.xml")
