+++ lending-rest-server/src/main/scala/info/armado/ausleihe/client/remote/LendingApplication.scala
@@ -1,0 +1,7 @@
+package info.armado.ausleihe.client.remote
+
+import javax.ws.rs.ApplicationPath
+import javax.ws.rs.core.Application
+
+@ApplicationPath("/rest")
+class LendingApplication extends Application
